Updated Function-Calling for Commands
Switched to using a dynamic import in events/commandinteraction.ts to
get rid of the need to manually add a case to the previously used
switch-case statement and having to import it as well, now working
automatically after updating commands.json (a.k.a. registering the
command in discord) and having the file created, now also having to make
sure to name the file the exact same as the command itself (e.g. if the
command is /do-something, the file must be named
commands/do-something.ts).