Restructuring: Config Files & Small Command Changes
Switched from using json files for the config and the slash commands
data to directly using typescript (.ts) files. Also added new property:
'commands' to the config, allowing for access to all commands' data from
within the code without having to add another parameter specifically for
this, which will be useful when making the 'help' command, as it allows
for searching through the code and dynamically creating the response.
Since the config and commands (.ts) files are now inside src/resources/
instead of directly being in src/, updated the imports for the 'Config'
type from src/config.ts to src/resources/config.ts. Also changed some of
the parameters for the 'embed' command, switching to using Attachments
instead of urls for the author and footer icons, as well as, in the
'webhook' command, removing the subcommand 'noavatar', removing the
option to have the webhook say something without specifying an avatar.